Rearrange rustcore to have a more logical file layout
